+++ client/src/components/WorkflowInvocationState/Export/index.js
@@ -1,0 +1,1 @@
+export { default as BioComputeObjectExportCard } from "./BioComputeObjectExportCard";
